admin: Adjust redirect logic to allow subdomain matching
The admin redirect address is determined by an heuristic which,
in brief, cross references the request's host header with the
receiving broker's advertised kafka addresses and looks up the
advertised kafka address of the leader node at the same index.

Previously the lookup was done by exact hostname match, but this
doesn't play nice with a k8s headless service. This commit adds
a fallback case to check whether an advertised kafka address is
appears to be a subdomain of the request hostname, by string
matching.

Includes a ducktape test for the basic 307 behavior.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>