producer_swarm.py: argument checking
If you accidentally pass in a float, which is very easy to do when
using `/` in python, _and_ the float has a fractional part, producer
swarm will fail to start. The service doesn't detect this in an obvious
way however, instead you will simply get a timeout at some point later.

So enforce that things that need to be ints are ints in the constructor.