r/consensus: fixed validating already matching batches last offset
When follower receives an append entries request that `prev_log_index` is
smaller than its own `prev_log_index` it validates if the
batches from the request matches (by checking a batch offset and
corresponding term) its own log. It that is the case the batches are
skipped to prevent truncation of valid batches and avoid the data loss.

The validation of already matching batches was broken if they happened
to be at the beginning of the log. In this case the `prev_log_index` is
not initialized being negative. This case was not correctly handled by
the logic calculating the next offset when checking matching batches.

That lead to a situation in which a range of batches starting with 0 was
never matching.

Fixed the issue by correctly adjusting the `prev_log_index` if it is
uninitialized.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 0d029cfd84fa9f2f0469fd2c5966768839cde075)