pandaproxy/sr: protobuf rendering: add test for explicitely set map_entry
Note that this is not considered to be a valid protobuf message.
However some thirdparty tools do generate such proto schemas
and we support it for compatibility with those.
For their support, we use a patched protobuf library on our side.

(cherry picked from commit 3abd347a3bc50a77596aaf3090ab270d99171793)