raft/tests: deflake truncation_detection_test
The current logic to wait on `enqueued` stage seems incorrect and is
causing the test to flake. `eneuqued` only guarantees that the request
is enqueued in the raft layer and subsequent replication may still fail
with not_leader error. So the sequence of flakiness is something like
this.

1. wait for enqueued futures to resolve
2. force a step down
3. wait for replicated entry to be truncated.

If (2) happens before the entry is appended to the leader log, the
replication future fails with a not leader resulting in a test retry.
Effectively (1) is not a tight enough condition to wait on to force
a step down.

This commit rewrites the test to replicate with a leader_ack and wait
for the future to be resolved (which guarantees a lleader append) and
then the replication monitor waiter should be resolved with a truncation
error after leadership change.

Additionally the test seems to be missing some cleanup between
retries, which is fixed

Signed-off-by: Michał Maślanka <michal@redpanda.com>