`storage`: use `segment_closed_exception` in adjacent func
In compaction related functions, we throw a `segment_closed_exception`
when encountering a segment that has been closed concurrently during
the process.

However, there is one location in `do_compact_adjacent_segments()` in
which we throw a `std::runtime_error` with a descriptive message of
the segment closed instead.

Change this throw site to use `segment_closed_exception`.