`compression`: add backwards compatibility check for `snappy`
The two committed files in `snappy_payload` are a raw uncompressed data file,
and a `snappy` compressed data file generated by `redpanda` using the
incorrect little-endian encoding for the version fields in the `snappy`
header.

They are used in a unit test to ensure that with the big-endian fix for
`snappy`, we are still able to decompress the buffer and get the same
decompressed data as before the fix.

(cherry picked from commit a84252d4b062bc95d7c70369cea22d42fbf5626b)