cpu-profile: tests for v2 format, stress test
Add some tests for the v2 format, and stress tests which try to
have something close to a worst-case scenario for CPU profiling, i.e.,
a large number of deep stacks.

In order to implement the stress test, we augment the stress fiber with
a mode which spins at a specific stack depth, i.e., calls N functions
before spinning so that the stack depth will be at least N (in practice
it will be somewhat more than N because there are other frames in the
stack beyond the ones added explicitly by the spinner).

Update existing DT tests for the new format.