bazel/toolchain: upgrade to clang-19
Upgrade to the latest clang. The real motivation here is that we want
the sanitizers to work, however for some reason the custom built clang
we use has false positives with coroutines and the sanitizers.

However clang-18 is compiled against a super old version of ubuntu and
missing libraries are missing (namely `libtinfo.5.so`) and not available
on recent ubuntu and fedora distributions.

Since we're upgrading the clang-19 we need to address a few deprecations
in libc++, so we upgrade seastar for this.