datalake: introduced function converting partition key to path
The value is converted to the string representation based on the transform
type and the key value. The path is constructed by concatenating the key
fields in the form of: <field_name>=<field_value> subsequent fields are
separated with '/'.

The value to string conversion is not described explicitly in the Iceberg
spec. The conversion is based on the reference implementation:

https://github.com/apache/iceberg/blob/fad0c1e6c68fbc7e48b5b17c02ed9c26a2693afb/api/src/main/java/org/apache/iceberg/PartitionSpec.java#L206

The conversion is done based on the following rules:
- Identity and truncate transform result values are rendered based on the
  source field type
- Bucket transform value is rendered as an integer
- Time transform values are rendered as a string representing date/time

Returned path elements are url encoded.

Signed-off-by: Michał Maślanka <michal@redpanda.com>