bazel/toolchains: use LTO compiled clang
Switch back to using a in-house compiled LLVM toolchain.

Due to the compiler being used on a variety of linux distributions
from CI to dev workstations, we want it to be as statically linked
as possible. These binaries here statically link the C++ standard
library and are compiled in ubuntu:jammy containers so it should have
quite a bit of compatibility.

Additionally, there seems to be issues with the sanitizers in the PGO
build, so we fallback to LTO only until we track down the issue.

Lastly, we switch to github releases from an s3 bucket. The s3 bucket
is in a non-ideal project and discoverability is hard. There are also
issues like being linked to a single AZ, etc so it's slow to download
in certain places in the world. We can punt all that to Github by using
Github releases to host the release artifacts from the containers in
`bazel/toolchain/Dockerfile.*`. We already have plenty of dependencies
on Github for building and if we want to fix that we can do it in a
unified approach to mirroring our dependencies and then this naturally
becomes apart of that solution.