r/consensus: switch back to main scheduling group in timeout_now
The `consensus::timeout_now` method triggers an immediate leader
election to handle the leadership transfers. Since the method is called
directly from the RPC request handler it is executed in the scheduling
group that the handler uses. In this case it is `raft` scheduling group.
This makes the whole leader election process to run in `raft` scheduling
group while it normally runs in `main`. Changed the code to always
handle timeout now in `main` scheduling group not to introduce
differences after the leadership has been transferred.

Signed-off-by: Michał Maślanka <michal@redpanda.com>