datalake: Add last_translated_timestamp to translation_stm
Intended to be a lower bound for the arrival time of the highest translated
offset. When we find untranslated records on a data source, we record the
current system time and the max translatable offset as a 'target'.
When translation meets or exceeds the cached target offset, replicate the
recorded timestamp. If the highest translated offset _exceeds_ the target, then
this will be an underestimate, which should allow us to _over_ estimate
instantaneous lag, hopefully improving our chances at achieving best-effort
translation deadlines.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>