cluster: implement a rate limiter for crash reporting
Introduce a new kvstore key space crash_tracker, a key in that key space
called rate_limiting_metadata with a serde value
crash_reporter_rate_limiting_metadata containing a timestamp of the last
time an attempt was made to upload crash telemetry data.

This is to provide a way of rate limiting crash telemetry uploads in a
way that is robust to crashes that may happen while uploading data.

This is integrated in the next commit to implement an upload rate of
1 request every 30s.