utils/fixed_string: support 0-length string, remove `const char*` cast
Cast to `const char*` was returning a pointer to an array without
terminating zero byte. This conversion is not in use and looks unsafe,
remove.

While certain compilers support them as extensions, 0-size arrays are
disallowed in C++. Clang is okay but ASAN is not happy about one.
Implement `fixed_string<0>` separately.