`rptest`: deflake `test_upgrade_java_compression`
This test expects that all segments will be compacted and re-written,
in order to recompress with the `snappy` compression fix post node restart.

However, this may not always occur for the test currently as it is written,
since some segments may be detected as already compacted post restart,
or not included in sliding window compaction due to already being cleanly
compacted.

To fix this, delete all the segment indices on disk to force a rebuild and
a self compaction of the segments.

Also, change the test to no longer wait on compaction, but on the consumer
succeeding instead.