`storage`: use `btree_multimap<>` for `ntp_by_gc_size`
Currently, if multiple `ntp`s were to have the same amount of data available
for reclaim, the entry in this map for that `size_t` key would be overwritten,
and only one `ntp` of the set would end up being garbage collected.

This is particularly bad since we are only considering the local retention,
and not the total available bytes that could be reclaimed due to data
being uploaded to tiered storage (i.e a common case here is
`usage.reclaim.retention == 0` for multiple `ntp`s, despite the fact that
resource management has identified data is available for reclaim from
tiered storage enabled topics).

Fix this by using a `btree_multimap` for the `ntp_by_gc_size` map to prevent
this collision case from occuring and blocking urgent garbage collection of
certain `ntp`s.