k/groups: do not throw from rebalance_timeout()
Do not throw an error when group::rebalance_timeout() is called on a
group with an empty members list.

This change makes the behaviour match Apache Kafka's behaviour in the
case when the last member of a group expires while there are pending
members in the group.

Currently the above scenario would lead to exceptions being thrown in
the join timer's callback and the group not being transitioned to the
Empty state on the expiry of the last member. Apache Kafka calculates
the rebalance timeout to be 0s when there are no members in the group
(only pending members) and transitions the group to the Empty state when
this 0s timeout expires and the join timer is triggered.

Ref: https://github.com/apache/kafka/blob/9570c67b8c4ed1a4c3888511adad58d9b3a8bc0f/core/src/main/scala/kafka/coordinator/group/GroupMetadata.scala#L409-L411
(cherry picked from commit cb8e50a5a7d404572deeaf0841bf7c17c1959961)