dt: test for expiry of last group member
This adds a regression test for the scenario when the last member of a
group expires while the group still has pending members.

It shows that the earlier log line for an exception in the timer
callback is no longer present. Furthermore, it shows that the group
correctly transitions to the Empty state the same way Apache Kafka does
in this scenario.