iceberg: support latest_protobuf_value mode
A new iceberg mode that also requires additional metadata (which
protobuf message to use), which means we no longer can store
iceberg_mode as a simple enum.

I converted it to a struct that is very similar API wise to the enum,
but for the protobuf case the API diverges.

This is a minimal change to add the new variant and update the API for
iceberg_mode. There are likely changes were we copy iceberg mode around,
and we might want to move, although TBH I expect the protobuf message
name to fit in the SSO optimization or be < 32 characters so I am not
sure the moves will help at all, but maybe references can be used
instead. I'll investigate updating that in another PR, although I expect
it to not be copied that much as this is really only used in setup for
the datalake and in admin-like requests.

(cherry picked from commit 2e754349603226e4fc32c817f352740d79f77efa)