datalake/sr: add a schema resolver based on latest subject version
Implements the proposal to work with protobuf messages that don't use
the sr wire format. Instead we trust the latest protobuf is
backwards compatible on the wire, then parse the record based on that.

We always assume the topic naming strategy. Right now there is no
caching of the schema, but we will add support for this in a followup
commit.

(cherry picked from commit 7b97bae67712296c14d714aea587053a8b3d4741)