cpu_profiler: add scheduler group
Add scheduler group to each sample. This is recorded on the seastar at
the moment of the sample side and we now include it in the output.

This is useful for understanding what is running in what scheduler
group.

Add a new unit test case for this functionality.

(cherry picked from commit 67a9d5383728e9fab3b732a3239aa53fab28d464)