rpk: enable serverless back for `rpk security user`
In 5e016b55 we introduced support for the public
API in `rpk security user`. However, we assumed
that serverless user were going to rely on this.

This was a bad assumption as serverless has a
special proxy for the API in Console and it's
already handled and added to the cloud profile.

So, a serverless cluster should create an Admin API
client and perform the requests as any other
self-hosted client would do.

(cherry picked from commit eca0ffc441d4381fbffa05d66244ae33af0473cd)