dl: do not enable translation stm when datalake is disabled
When `iceberg_enabled` cluster configuration property is set to false we
do not have to create `datalake::translation::state_machine` for each
partition. This way when datalake capabilities are disabled in the
cluster we will not have to reply the whole log for the datalake state
machine.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 485e9a4fff873bf9d64489f47580fa9581006ebc)