r/stm_manager: do not populate batch cache when background apply happens
When `raft::state_machine_manager` executes the background fiber apply
it reads batches that are not at the tip of the log as for the
background fiber to start the STM must be out of date.

Previously we populated the `batch_cache` with one hit wonders coming
from state machine catching up. This increased the CPU usage a lot when
state machines were executing `apply` logic. For the large topics the main
most of the CPU cycles are burned in the batch cache reclaimer.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 22aa0a0382071c10d6b676ac483324d5fc958cf8)