dl/translation: add ability to translate from the middle of a batch
Translation can be aborted in the middle of a batch which is then
flushed to the coordinator and the subsequent iteration should start
from the later offset. Since log reader only operates on batch
boundaries we need to layer this additional check in the multiplexer to
skip already translated offsets (by passing an additional param).

(cherry picked from commit 6dc1950bc71096b80a6eae792c2055630602f595)