dt/redpanda: send signal to threads more reliably
This fixes the flakiness in the signal-based crash loop tracking tests.
The tests were occasionally failing (~3/100 times locally) when the
signal sent to a redpanda thread for testing was handled by a different
shard than expected.

The current implementation for sending signals to specific threads is
unreliable because it uses `kill`, which targets the entire process
group rather than a specific thread. The reason why this works most of
the time today is because of the way the `kill` syscall selects a thread
to wake up after it has set the signal to pending for the process group.
The syscall completes by attempting to wake up a task from the process
group to handle the signal, and during its selection, it prefers the
thread that was specified in the syscall. This means that in the
majority of the cases when we pass a thread id to the `kill` syscall, it
ends up being handled on the target thread. However, this is not
guaranteed and depends on the state of the threads.

To ensure signals are reliably delivered to the intended thread, this
commit switches to using the `tgkill` syscall, which sets the signal to
pending for the target thread only. Therefore, it guarantees that the
signal is handled on the target thread.

Ref:
- https://elixir.bootlin.com/linux/v6.13.1/source/kernel/signal.c#L937-L1034
- https://man7.org/linux/man-pages/man2/tkill.2.html
- https://man7.org/linux/man-pages/man2/kill.2.html

(cherry picked from commit 8c662a77555eebd8e13cb6456c1e54e4cb850e9d)