archival: Replica validator and allow-gaps prop.
Previously, when the uploads were paused and gaps are allowed the resume
cold be blocked by the replica_state_validator. If the leadership
transfer happens during the pause and the gaps are already created the
validator would prevent uploads from progressing without taking into
account 'allow-gaps' property.

This commit fixes the problem by checking the property when the
validator output is examined. If the validator detects anomaly but the
only anomaly is a gap anomaly and the 'allow-gaps' is set to 'true' the
upload loop is not blocked.

This allows 'ntp_archiver' to resume uploads properly.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>