bytes/details: add io_fragment::unoptimized_release method
The usual `io_fragment::release` method trims and potentially allocates
a new temporary buffer to release. This results in a behavior that is
different from `io_fragment::share`. This new
`io_fragment::unoptimized_release` can act as a drop-in replacement for
`io_fragment::share` in cases where ownership can be moved. This further
avoids unneeded reference counting in these cases.