crash_tracker: handle exception while initializing
Previously if there were an exception while trying to initialize the
crash tracker, that would lead to an assertion failure. This is not a
great UX.

This commit changes the behaviour to skip writing a crash report if the
crash tracker is not yet initialized.

A test is added to simulate this described behaviour.