kafka/server: address issues in consumer group metrics test
Parameterized value was being modified during the test. This messed
up the reporting during the test. It is now copied to a local variable.

Always using the same group name made the test fragile. A hanging
consumer on the first run would be picked up by consecutive reruns.
Randomized the group name on every run.

Added error message in create_consumer function to be able to provide
some context.