`storage`: add bytes before calling `release_appender()`
There still seem to be races with the removal of segments that have been
"closed" (i.e don't have an appender) and the addition/subtraction of
their bytes.

For example, it is possible that `release_appender()` is called,
a segment's appender is released, and a concurrent race with prefix truncation
results in a segment's bytes being subtracted before they are added in
the first place.

Fix this issue by adding the bytes before calling into `release_appender()`
in order to guarantee that any bytes removed from a segment without an
appender will have been added previously.

(cherry picked from commit 772eb61276593d147e039b09081bfa602ab71458)