cpu_profiler_test: do not customize CPU count
We hardcoded CPUs to 10, but any custom CPU number is fraught because
it has to work in an "unknown" CDT environment, docker CI environment
and localhost docker environment (which is not the same as CI as the
number of CPUs, for example is exposed from the host to the container).

So stop trying to do this and just make it a default resources test as
the other ones in this file.

Also, use 1 redpanda node for all of these tests as I see no benefit in
using a 3 node cluster.

(cherry picked from commit 2181cba35cff84d709d7b8ace6421b84bbe0eea6)