`cluster`: use `node_id` in `self-test-prefix`
If the `cloudcheck` fails and is retried, it is possible that the number
of objects uploaded to the shared `self-test/` folder in cloud storage
will exceed the number of objects listed by a test call to `list_objects()`.

This could cause the test to fail with `Uploaded key/payload could not be
found in cloud storage item list`.

To prevent interference between nodes in the cluster all running the
`cloudcheck`, use the `node_id` as part of the prefix for the `self-test`
folder.

(cherry picked from commit 1a19a4beee8b75b3026cc4cccf34e48252bde86f)