s3: Add timeout to credentials fetching operation
The s3_client uses credentials propagated by the
'cloud_storage::remote'. In case if IAM-roles are used the credentials
are fetched by the background operation which is started right after
startup.

Normally, this is not a lengthy process. The credentials are acquired
within first few hundered milliseconds after Redpanda starts. The client
pool performs self configuration while holding units from the
'_self_config_barier' semaphore.

If the credentials can't be acquired the 'auth_refresh_bg_op' will log
it and will never create a 'cloud_roles::refresh_credentials' instances.
This meanst that the credentials will never be populated and any attempt
to start upload will stuck waiting for self configuration.

The problem is that in this case Redpanda fails to stop gracefully. The
'ntp_archiver' attempts to upload the manifest right before the
leadership transfer. Because of that if the 's3_client_pool' is stuck
waiting for self configuration to finish and at the same time node
decomissioning starts the node will never decomission correctly because
it will fail to stop partition.

When the 'cluster::partition' is stopped it invokes 'stop' method of the
'ntp_archiver'. Prior to that it invokes graceful leadership transfer
method which forces 'ntp_archiver' to upload the manifest. This
operation is getting stuck indefinitely because of the self
configuration problem. The graceful leadership transfer in the
'ntp_archiver' is time bound. After 5s interval it will give up and the
'cluster::partition::stop' method will be invoked. It will invoke
'ntp_archiver::stop' method which in turn will try to close the gate.
But the gate will be held indefinitely by the manifest upload which got
stuck. As the result the 'partition_manager' will not be able to stop
the partition.

This commit fixes the problem by adding a timeout to the acquisition of
units from the '_self_config_barier'. With this change the manifest
upload will eventually timeout and release the gate.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>
(cherry picked from commit 35a8280904282a60571eebafb75ec2169568098f)