dl/throttle: renamed the backlog_status field
Renamed the backlog status field to better indicate the meaning of it.
Previously the name was indicating that the field was related with
translation backlog while in reality it is the number of partitions that
can not keep up with the translation.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 582077034b4172d7994fce9b481db1454966d5c2)