rptest: Fix single partition test
The test is flaky because we added a new command. The test is trying to
estimate the delta offset based on number of uploaded segments. The
archiver adds one metadata record per segment. But also it adds one
'mark_clean' command and one 'update_highest_producer_id' command.

So the test assumes that if the througput is high the 'ntp_archiver' is
uploading 4 segments in every iteration. This means that all these extra
metadata records are added only once per iteration. Which is just two
extra metadata records per four segment metadata records. Which gives us
a value of 1.5 for the delta estimation.

If the throughput is not at the top the 'ntp_archiver' makes less than
four segment uploads per iteration. In this case the metadata overhead
is larger. The max overhead is two metadata records per every segment
record. So the delta is number of segments multiplied by 3.

But because the additional command was introduced (rw-fence) this
reasoning no longer holds. In order to estimate the bounds for the delta
offset we need to multiply number of segments by 1.75 and 4.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>