datalake: expose translation disk bytes flushed
This reflects how much data as translator has flushed to disk since the
last time that it "finished" by uploading and removal local staging
data. There is a brief moment during finishing where this metric will
return zero, but the data is still on disk. However, in this case it is
a short lived situation because the translator context is in the process
of uploading and deleting the data.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit 57c0e213d860f896cfe4a4b0faa2c9761ba8d41f)