r/consensus: adjust learner initial offset to batch boundary
Redpanda Raft protocol implementation operates on batches. The batch is
the smallest unit leader can send in append entries request to the
followers. The fact that the batch is the smallest unit Raft operates at
forces the snapshots to be taken at batch boundaries. In other words the
last snapshot included offset must be a last offset of a batch.

Described invariant may not always be true when controller backend
request fast reconfiguration and the offset provided is bounded by
max_collectible_offset. The `max_collectible_offset` may be in the
middle of the batch. (an example of state machine returning offset in
the middle of the batch is datalake translation state machine which
operates at the record level).

Added code that will adjust the requested learner start offset to batch
boundary. The adjusted offset may not the first possible batch last
offset because it uses index lower bound. (Index is sparse and using
32 KiB steps)

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 9659442455d8129db0c12d56526c628a9b2eb0dd)