archival: Add concurrency control to segment_collector
It is possible for the compaction to race with segment lock acquisition.
The 'collect' method is not asynchronous but subsequent call to
'make_upload_candidate' is. The method acquires locks asynchronously. So
if this process races with compaction the acutal content of the segments
would become unexpected. In order to avoid this we need to check that
the expectation holds after all segment locks are acquired.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>
(cherry picked from commit accee8af23d053722db55a60f5483d4cd67791fc)