datalake: add config for max translator flushed bytes
Forces a translator to finish if flushed bytes exceed the configured
value. The hard-coded value of 32 MB seems reasonable, but there are two
reasons for making it a tunable (ie non-user). First, the mechanism of
the limit existing is a safety net until we have confidence in space
management's ability to control usage, so allow it to be tuned since it
is difficult to predict what might be useful to change. Second, tuning
this value in the context of tests is necessary in order allow lots of
data to accumulate and trigger finishing translators.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit 652333a5cae3a7ab5b82b9fd325d41a3c8a903e4)