datalake: add translator finish_translation interface
Adds an interface to request a translator to finish translation even
when the translator is idle (neither running nor waiting). This is
important for making a translator responsive in low disk usage
situations since meaningful amounts of on-disk data may be under
management of a translator and stop_translation--the only current
preemption mechanism--is designed to stop translations that are in the
running state.

Ideal finish_translation is merged with a more generic version of
stop_translation, but that is a bigger project (in the context of the
partition_translator implementation) because the life cycle of
start/stop translation interfaces is tightly coupled with the scheduler
(e.g. intrusive list membership, expecte statuses and notification of
state changes).

More details:

The partition translator operates in a nested structure. At the outer
layer is a polling loop that waits for new offsets to translate:

   translate_until_stopped
      poll(fetch_translation_offsets)

once new offsets are available then the next layer waits for the
scheduler to allow it to run and process the offsets:

   translate_until_stopped
      poll(fetch_translation_offsets)
         run_one_translation_iteration

Notice that there are two places where control can block: at the outer
later if there are no new offsets to process (idle topic) and in the
nested level waiting for the scheduler to schedule the translation.
Currently the translation interface stop_translation is only capable of
preempting a running translation. That is, one which is executing at the
nested level.

This is problematic when it comes to controlling the disk space usage of
a translator because we need to take into account the usage of idle
translators which are neither running or waiting to be schedule. They
may have a lot of staging data, and need to run shortly to finish up the
current translation. In this case they are likely polling in the outer
layer waiting for offsets. However, currently the only way to preempt
this outer layer polling is by completely shutting down the translator
using an abort source. That is not what we want to do since we'd have to
rebuild the translator.

What this commit does is add a new translator interface,
finish_translation which signals to the translator to finish, and can be
used when the translator is in any state. The partition translator
implements the interface by setting a flag which the outer polling layer
examines in order to break out of its polling loop even if there are no
new offsets to translate.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit b29130763de0299425f8ccab247ba4666d5eb8e6)