log: added method allowing to find indexed batch base offset upper bound
Added `storage::log::index_batch_base_offset_lower_bound` a method
returning a `base_offset` of a batch that is smaller than or equal to
the requested offset. The method uses segment index to find the offset
so returned value may not be the first possible batch with smaller
offset but the previous one indexed.

Signed-off-by: Michał Maślanka <michal@redpanda.com>