datalake: add scheduler interface for immediate finish
This commit introduces a new interface:

   scheduler::request_immediate_finish(vector<translator>)

which can be used by an entity external to the scheduler for requesting
that a set of translators be promptly stopped and finished. The primary
use case for this interface are global resource schedulers (e.g. disk
space) which want to inject priority scheduling decisions. Translators
request for immediate finish should promptly upload and remove their
staging data from disk.

This commit adds the basic scheduling component, and the next commit
sprinkles around some changes that make scheduling more responsive.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>