dl/translation: use total backlog in datalake backlog controller
The total backlog is the value that should be use by backlog controller
as it is directly proportional to the amount of work to be done during
translation. Change the backlog controller to use the total backlog
instead of per partition average and adjusted controller defaults.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 3247345774898d9ca96469d9568cbfb594db9291)