datalake: ignore recoverable errors on writer flush
Recoverable errors are things that might happen while writing like
running out of memory or timing out. In both of these cases the state of
the writers is consistent, but the translator should flush and possibly
finish its current translation batch.

The problem prior to this commit was that when flushing the local
parquet file writer would treat recoverable errors as non-recoverable
errors. So in a case like oom the translator would try to flush prior to
finishing and encounter an unexpected error like this:

   node-0: WARN  2025-04-09 16:00:45,750 [shard 0:data] datalake -
   {kafka/foo/0}-term-1 - partition_translator.cc:308 - Translation attempt
   ran into an unexpected exception: std::runtime_error (Error flushing
   in-progress translation: writer flush error)

In response the entire translation would be thrown away and restarted.
Depending on why the translator was interrupted, it's possible that this
would turn into an infinite loop without the translator being able to
make progress.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit de9f0e120dfed7f1384f3aefadebfd713ce786f3)