datalake: propogate disk usage error instead of oom
Adds an out_of_disk reason for stopping a translator. This is plumbed
down into the partition translator where it has the same affect as an
oom error, namely to finish immediately.

This commit swaps out the hard coded oom reason in finish_translation
for out of disk reason, as the only current user is disk space
management.

In the next commit we will plumb the reason for finish requests from
their source, because finish_translation is generic and not specific to
an out of disk reason.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit 0d4a62a9c32817f5e055bc3f8ad28119011f2979)