datalake: check for recoverable error when finishing
Prior to this change we did not check if the local parquet file writer
was storing a recoverable error which meant that in cases like oom or
out of disk that a recoverable error (instead of results) would be
propogated back to the caller. the caller would in turn handle the
recoverable error but wouldn't have any results to finish.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>