r/stm_manager: apply snapshot in background apply fiber if required
The background apply fiber is used in state machine manger to catch up
the state machines which are lagging (f.e. state machine apply method
thrown an exception). The background apply fiber was missing the
snapshot application. This lead to a rare situation in which a snapshot
delivery while the state machine was being caught up in background apply
might have lead to situation in which a reader can not continue as the
next offset was not updated and below the raft start offset.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit f4a4af69c546fd74344dc3b3016afb556c731a36)