rpk: infer SR address from other listeners
We had bi-directional inference of address:port
between the Kafka and Admin API listeners in rpk.
Now, we've added support for the Schema Registry
in both the `rpk` section of redpanda.yaml and in
the rpk profile (rpk.yaml).

This commit introduces logic to infer SR (or Kafka /
Admin API) listeners based on whichever listener is
available.

We chose to prioritize inferring in the order:
  Kafka > Admin > SR

Once we infer one listener, we don't look at the
rest. This isn't perfect or fail-proof, but it's a
reasonable fallback until we support full
bi-directional port discovery between listeners in
Redpanda.