`storage`: check for exceptions in `log_manager::disk_usage()`
A single log should not be able to block space management of all local
logs.

To avoid this existing behavior, inspect the result of `log->disk_usage()`
in a `then_wrapped()` continuation within `log_manager::disk_usage()`.
We will log a message and return an empty usage report if the disk usage
call failed for any reason, but will allow the log manager to continue on
checking the other logs on this shard.