`storage`: handle exception in `disk_log_impl::disk_usage`
In the case that `disk_usage_and_reclaimable_bytes()` throws an
exception for any reason, return the usage as the persisted size on disk
with `0` bytes indicated as reclaimable.

This will help space management take into account the space used by a
partition that is throwing an exception when calculating its disk
usage statistics (even if none of it is marked as reclaimable).