storage: s/max_collectible_offset/max_removable_local_log_offset
With the new pinned offset, it's not immediately clear from the name how
the two differ. So this takes a stab by changing max_collectible_offset
to be max_removable_local_log_offset, indicating 1) it is an offset in
the local log, and 2) it is a log offset rather than a data offset.

Along the way, replaces "collectible" terminology with "removable" -- it
was never really clear to me what "collectible" referred to, but
"removable" seems to capture the fact that the records can be removed
(either by retention or by compaction).

(cherry picked from commit b77eaf0deeedd9ff7ec3db32704b8da34c3152fd)