tests/write_caching: limit number of messages sent in write caching test
Write caching tests expects a data loss to happen. In order for this to
be true some messages must be kept in memory and not persisted to the
disk. Any disk flush will prevent a data loss event from happening.

In `recovery_stm` we have a policy that flushes every 1 MiB of data to
make sure that the Raft commit index progresses incrementally during
recovery.

The test failed intermittently when a recovery of the follower happened
and the follower log was flushed. Changed the test to use less messages
not to hit the `recovery_stm` checkpoint threshold.

Signed-off-by: Michał Maślanka <michal@redpanda.com>