datalake: add disk manager interface
The disk manager interface is what schedulers use to interact with the
disk manager on core 0 when requesting disk reservation units.

In this commit the interface is added, and plumbed into the schedulers,
not the final implementation is not defined within the datalake manager
because it involves references to scheduler interfaces that have not yet
been introduced.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit d0b42c9965b8374ed345534fb22a8c9d4f01c893)