datalake: implement core 0 disk reservations
Adds an implementation of reserve_disk for core 0 disk manager. This
interface is invoked by schedulers when they run out of disk space
reservation. Core 0 will try to give the scheduler a block reservation.

If providing the reservation causes core 0 to hit the soft limit, then
it kicks the disk manager control loop to further deal with the
situation by attempting to acquire units to bring usage back below the
soft limit.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>