datalake: core 0 monitor should harvest unused reservations
Before core 0 requests translators to finish, it first should check to
see if the reason it is below the soft limit is because schedulers are
holding on to unused disk reservation. It collects all the unused and
then reevaluates the soft limit.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit 86f5d062d9636f985abd9447c5d834374c354ed8)