`storage`: add `_compaction_placeholder_enabled` to reducer
Previously we would gate `make_placeholder_batch()` in `copy_data_segment_reducer`
based on some logic with the assigned `_segment_last_offset`.

This comparision of offsets isn't a very strong guarantee- the better thing
to do is to construct the `copy_data_segment_reducer` with the boolean
instead, and use that within `filter()` to ensure a placeholder batch is
installed only if the `feature::compaction_placeholder_batch` feature is enabled.