`storage`: add `storage::internal::should_keep()`
Encapsulate the logic used in two places for determining if the last
record in a segment needs to be persisted in a new function in
`segment_utils.h`.

Also refactor the two call sites to `copy_data_segment_reducer`
within `segment_deduplication_utils.cc` and `segment_utils.cc` so that
they are more similar in style, to improve readability.

This commit also has a functional change in that self-compaction will
ensure the last record in a segment is persisted, whereas before, if
the `compaction_placeholder_batch` feature was disabled and the last
record in a segment was a tombstone, the record _would_ be removed.
This went against the invariant we have upheld to now in which we
ensure segments cannot be empty of batches.

(cherry picked from commit 2d6daa6e2006d7ba12f6811f0db33073c2babeaf)