crash_tracker: improve diagnostics on backtrace failure
If the stack is corrupted by the time the signal handler runs,
a regular (non-incremental) backtrace call may segfault.
This makes it difficult to diagnose the origin of the signal.

This commit adds a functionality to use seastar's incremental backtrace
method to print the backtrace immediately after entering the signal
handler.

This behavior is gated behind the log level of the crash_tracker logger.
By default, this is disabled to avoid printing the backtrace twice—
once here, and again by Seastar's signal handler.