rpk: prevent security role in cloud
This was mistakenly assuming that it was not
available in serverless, but available for BYOC
and dedicated.

However, this command space uses the admin API
and should not be run against Redpanda Cloud atm.
This commit adds the check so it exits with a
better error message in case you run it from a
cloud profile.

(cherry picked from commit b2c8ba35ad7f1e67e8eda59d3e86df9d23edeabc)