config: Introduce 'ignore_existing_node_id' to node_id_override
Originally, the intended usage here revolved around situations where a blank
node was unable to join the cluster due to quorum loss caused by "ghost nodes".

Now, we would like to add the ability to force a known UUID:ID pair into a
node which is _already_ a member of the cluster (i.e. has a node ID already).
To distinguish this new usage from the original usage, we introduce
'node_id_overide::ignore_existing_node_id'.

For a YAML based config, we add a boolean 'ignore_existing_node_id' field to the
existing YAML map.

For a CLI based config, we introduce a '/ignore_existing_node_id' suffix for the
node-id-overrides option, as follows:

- regular usage (as before): '<uuid>:<uuid>:<id>'
- force usage: '<uuid>:<uuid>:<id>/ignore_existing_node_id'

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>