`storage`: add check for closed segment to `index_chunk_of_segment_for_map()`
We weren't checking whether the `segment` to index was closed or not before
obtaining its read lock and trying to perform I/O on it.

Fix this by checking if the `segment->is_closed()` and throwing an exception
if it is before proceeding.

(cherry picked from commit 395ec1f87ed0b13807aadf80d84a47a9ce2ff317)