`storage`: add `single_segment_reader`
Meant for reading batches from a single `segment`.

This does not consider any other `log` state- meaning, the `log`
could have been prefix truncated and batches read from the `segment`
could be ahead of the new `_start_offset`.

This class should only be used when it is necessary to consider the
batches still physically stored in a `segment` on disk, e.g for file
position calculations. For all other applications, the `log_reader`
should be used.