`storage`: use `single_segment_reader` in `get_file_offset()`
This function could run into issues with the log `_start_offset` when
making a `log_reader`.

Consider the case in which:
1. A prefix truncation intersects a `segment`, which bumps `_start_offset`
   but doesn't alter the `segment` or `segment_index` on file.
2. A call to `offset_range_size()` is issued.

Currently, we are either assuming the found entry from a `segment_index`
has a `base_offset >= _start_offset` AND we're assuming that `segment`
passed has `s->offsets().get_base_offset() >= _start_offset`, neither of
which may be true due to the prefix truncation. This can lead to a thrown
exception in the `log_reader`: `std::runtime_error: Reader cannot read
before start`.

Fix the problem by using a `single_segment_reader` (which reads the physical
batches on disk, not subject to the same constraints of `log` state) instead
of the `log_reader` to avoid the problem of having to read before the new
`_start_offset`.