c/health_monitor: include offline partitions in leaderless info
When all the nodes hosting partition replicas are offline their health
reports are not available. This leads to a situation in which a
partition which is clearly leaderless (there are no online replicas so
there must be no leader) is not included into the cluster health
overview.

Fixed that issue by detecting the offline partitions and accounting them
as leaderless. The detection is based on the topics table as the health
report for the offline node is unavailable.

Signed-off-by: Michał Maślanka <michal@redpanda.com>