cluster: add `topic_table::get_name_by_id(...)`
Implement an efficient lookup for finding a topic by its topic id. This
will be used by kafka handlers to translate the topic id received on a
request to its topic name.

The optional topic id is always checked for nullopt-ness, since the code
needs to work during the migration period when not all topics have topic
ids assigned.

This is implemented using a wrapper class `underlying_map` as well as by
making various fields of `topic_metadata_item` private to provide safer
access to the value type.

The goal is to prevent accidental mutation of the `tp_ns` and `tp_id`
inside the `topic_metadata_item` as much as possible, since that would
violate the invariants of the `by_tp` and `by_id` maps.