`storage`: use `std::deque` as `segment_set::underlying_t`
`ss::circular_buffer` allocates contiguous chunks of memory for its
storage. We have seen instances of oversized allocations and even worse
crashes due to failures to allocate memory in the `storage` layer due
to the use of it as the underlying container in the `segment_set`.

Historically, it was swapped in (in commit `b5f477c`) for `std::deque`
due to `gcc`'s lack of a `noexcept` constructor. Since `gcc` is no longer
used to build `redpanda`, this is no longer a constraint (`clang`'s `libc++`
does the right thing and supports `noexcept` constructors for `std::deque`).

Revert use back to `std::deque`, which allocates non-contiguous chunks
for its storage, allowing us to avoid oversized allocations and potential
crashes for large amounts of `segment`s.