`config`: add `log_compaction_merge_max_segments_per_range`
A tunable property that allow control of the maximum number of `segment`s
in a range which can be concatenated to form a single `segment` during
an adjacent merge operation.

A value below 2 effectively disables adjacent merge compaction, a value of
2 has similar behavior to the previous adjacent merge implementation, and
a value of `std::nullopt` imposes no maximum on the number of `segment`s
which can be combined at once (although raft term and
`max_compacted_log_segment_size` still impose restrictions on the potential
size of an adjacent compaction range).