serde/json: add a json writer
It handles fragmented memory, writing large strings, but doesn't try to
be efficent, which I think is mostly fine. The trickest part is writing
valid utf-8 strings. I tried to write a bunch of tests to validate the
behavior.

Also note that numbers are only written with 6 digits of precision. I
don't really think we care about more than that.