k/client: caller-side malformed response handling
This shifts the responsibility for detecting malformed responses to the
caller. While these responses are not expected in normal operation, it's
safer to explicitly check and throw an exception than risk undefined
behavior.

We’re making this change now because upcoming changes to the
list_offsets internals will make it harder to perform this validation
inside the client.

As part of this, the thrown exception type has been changed from
kafka::client::broker_error to kafka::exception. This is safe, because
both inherit from kafka::exception, so the SR HTTP responses continue
to surface the error code and message back to the client as before.

Ref: https://github.com/redpanda-data/redpanda/blob/bfafcdf396b8096546d42a44f90a50852c3fcda8/src/v/pandaproxy/reply.h#L152-L167