tests/Dockerfile: add docker cache for apt
Add a --mount=type=cache for apt downloaded package and lists, which
speeds up the image build by a significant amount. The apt install
process works as normal, but using the cache allows it to carry over
downloaded packages both from one stage to another in the same build,
and also between builds.

The list files in /var/lib/apt/lists are also cached, which means that
apt update is also much faster on subsequent builds, as it does not need
to redownload package index files unless they have changed.

This shaves 20-30% off the full rebuild time for me (e.g., 10 minutes
to 7 minutes), when the apt cache is populated. The image size is
unchanged at 13.1 GB.